Split global handers into files
