+++ internal/pkg/keybindings/global_help.go
@@ -1,0 +1,42 @@
+package keybindings
+
+import (
+	"github.com/awesome-gocui/gocui"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
+)
+
+type HelpHandler struct {
+	GlobalHandler
+	ShowHelp *bool
+}
+
+func NewHelpHandler(showHelp *bool) *HelpHandler {
+	handler := &HelpHandler{
+		ShowHelp: showHelp,
+	}
+	handler.id = HandlerIDHelp
+	return handler
+}
+
+func (h HelpHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		tmp := toggle(*h.ShowHelp)
+		h.ShowHelp = &tmp // memory leak?
+
+		// If we're up and running clear and redraw the view
+		// if w.g != nil {
+		if *h.ShowHelp {
+			v, err := g.SetView("helppopup", 1, 1, 145, 45, 0)
+			g.SetCurrentView("helppopup")
+			if err != nil && err != gocui.ErrUnknownView {
+				panic(err)
+			}
+			keyBindings := GetKeyBindingsAsStrings()
+			views.DrawHelp(keyBindings, v)
+		} else {
+			g.DeleteView("helppopup")
+			g.SetCurrentView("listWidget")
+		}
+		return nil
+	}
+}
